Refactor: Move CardListTile to components/
